fix: ensure that linaria plugin to run first
fixes #333
