+++ src/babel/utils/toValidCSSIdentifier.js
@@ -1,0 +1,5 @@
+/* @flow */
+
+export default function toValidCSSIdentifier(s: string) {
+  return s.replace(/[^_0-9a-z]/gi, '_');
+}
